Fix usability bugs in GUI. Can edit SC2Banks outside of SC2 dirs.
Don't empty model upon save/overwrite. This is more intuitive --
before it looked lik the user lost the data. The only hint that the
information was saved is that the Recorded Signature is updated
sanely.

Rename some of the UI elements -- Old Signature is now Recorded
Signature; Update Signature is now Overwrite Signature. This makes
their intent clearer.

Don't blow up when the Model is incomplete and the UI calls
Model.calculate_signature.

TODO: Add tests!